feat: cleanup, apply game dashboard layout
