+++ client/src/router.ts
@@ -20,7 +20,6 @@ import Manual from "@port-of-mars/client/views/Manual.vue";
 import Home from "@port-of-mars/client/views/Home.vue";
 import Privacy from "@port-of-mars/client/views/Privacy.vue";
 import Profile from "@port-of-mars/client/views/Profile.vue";
-import Solayout from "@port-of-mars/client/views/Solayout.vue";
 import store from "@port-of-mars/client/store";
 import {
   ADMIN_PAGE,
@@ -38,7 +37,6 @@ import {
   ABOUT_PAGE,
   PRIVACY_PAGE,
   PROFILE_PAGE,
-  SOLAYOUT_PAGE,
 } from "@port-of-mars/shared/routes";
 
 Vue.use(VueRouter);
@@ -86,7 +84,6 @@ const router = new VueRouter({
     { ...PAGE_META[ABOUT_PAGE], component: Home },
     { ...PAGE_META[PRIVACY_PAGE], component: Privacy },
     { ...PAGE_META[PROFILE_PAGE], component: Profile },
-    { ...PAGE_META[SOLAYOUT_PAGE], component: Solayout },
   ],
 });
