+++ shared/src/routes.ts
@@ -12,7 +12,6 @@ export const HOME_PAGE = "Home" as const;
 export const ABOUT_PAGE = "About" as const;
 export const PRIVACY_PAGE = "Privacy" as const;
 export const PROFILE_PAGE = "Profile" as const;
-export const SOLAYOUT_PAGE = "Solayout" as const;
 
 export type Page =
   | "Admin"
@@ -28,7 +27,6 @@ export type Page =
   | "Profile"
   | "Verify"
   | "Manual"
-  | "Solayout"
   | "Privacy";
 
 export const PAGES: Array<Page> = [
@@ -46,7 +44,6 @@ export const PAGES: Array<Page> = [
   HOME_PAGE,
   ABOUT_PAGE,
   PRIVACY_PAGE,
-  SOLAYOUT_PAGE,
 ];
 
 export function isPage(pageName: string): pageName is Page {
@@ -164,13 +161,6 @@ export const PAGE_META: {
     name: PROFILE_PAGE,
     meta: {
       requiresAuth: true,
-    },
-  },
-  [SOLAYOUT_PAGE]: {
-    path: "/solayout",
-    name: SOLAYOUT_PAGE,
-    meta: {
-      requiresAuth: false,
     },
   },
 };
